pyproject.toml
[project]
name = "markdowner"
version = "0.1.0"
description = "Clipboard OCR to Markdown using a local Hugging Face VLM (Qwen2.5-VL by default)"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "transformers==4.47.1",
    "accelerate>=0.33.0",
    "pillow>=10.2.0",
    "huggingface_hub>=0.24.0",
    "markupsafe==2.1.5",
    "torch==2.6.0",
    "bitsandbytes==0.47.0",
    "torchvision>=0.21.0",
    "pynput>=1.8.1",
    "pyperclip>=1.11.0",
    "addict>=2.4.0",
    "matplotlib>=3.10.7",
    "einops>=0.8.1",
    "easydict>=1.13",
]

[project.optional-dependencies]
notebook = [
    "notebook>=7.1.3",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
